feat/868: moved _synthesizeWithSpecificVendor to tts-task.js
